twittersummarization/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
cffi==1.11.5
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.0.9
enhance==1.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
langdetect==1.0.7
lxml==4.2.5
parsel==1.5.1
py-translator==1.8.9
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.7.2
pyOpenSSL==18.0.0
PySastrawi==1.2.0
python-slugify==1.2.6
pytz==2018.7
queuelib==1.5.0
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
Unidecode==1.0.22
w3lib==1.19.0
zope.interface==4.6.0
